+++ bindings/nostr-sdk-js/src/relay/mod.rs
@@ -7,12 +7,10 @@ use std::ops::Deref;
 use nostr_sdk::prelude::*;
 use wasm_bindgen::prelude::*;
 
-pub mod filtering;
 pub mod flags;
 pub mod limits;
 pub mod options;
 
-use self::filtering::JsRelayFiltering;
 use self::flags::JsAtomicRelayServiceFlags;
 use self::options::{JsRelayOptions, JsReqExitPolicy, JsSubscribeOptions, JsSyncOptions};
 use crate::database::JsEvents;
@@ -124,11 +122,6 @@ impl JsRelay {
     /// Get Relay Service Flags
     pub fn flags(&self) -> JsAtomicRelayServiceFlags {
         self.inner.flags().clone().into()
-    }
-
-    /// Get relay filtering
-    pub fn filtering(&self) -> JsRelayFiltering {
-        self.inner.filtering().clone().into()
     }
 
     /// Check if relay is connected